+++ deno/v0.1/build-parsley.sh
@@ -6,8 +6,8 @@ ts_config=$curr_dir/tsconfig.json
 parsley=$curr_dir/deno/parsley.ts
 parsley_test=$curr_dir/deno/parsley.test.ts
 
-parsley_es=$curr_dir/esmodules/parsley.js
-parsley_test_es=$curr_dir/esmodules/parsley.test.js
+parsley_es=$curr_dir/../../es/v0.1/parsley.js
+parsley_test_es=$curr_dir/../../es/v0.1parsley.test.js
 
 # deno cache --reload ./deno/parsley.ts
