+++ deno/v0.1/parsley.ts
@@ -43,4 +43,4 @@ export {
     // createChunkArrayInjection,
     // createNode,
     // createTextNode,
-} from "./structure/structure.ts"
+} from "./fragment_builder/structure.ts"
