+++ deno/v0.1/type_flyweight/hooks.ts
