+++ deno/v0.1/type_flyweight/template.ts
