+++ deno/v0.1/type_flyweight/text_vector.ts
