--- v0.1/deno/fragment/fragment.ts
@@ -1,6 +1,0 @@
-interface Fragment<N, A, P> {
-    connect(): void;
-    update(params: P): void;
-    disconnect(): void;
-}
-
